style: PlayerName -> CarName 으로 클래스명 수정
