+++ README.md
@@ -33,7 +33,7 @@
 - 차가 이동 가능한지 여부를 제공 
 
 
-### PlayerName
+### CarName
 
 - 5글자 이하의 이름만 허용한다
